consequnce-backend/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.45.0
django-cors-headers==3.7.0
django-rest-passwordreset==1.2.0
django-sendgrid-v5==1.1.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-flex-fields==0.9.2
idna==3.2
mysqlclient==2.0.3
oauthlib==3.1.1
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.19.0
python-http-client==3.3.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
sendgrid==6.8.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
urllib3==1.26.6
